\name{interlabstats}
\alias{interlabstats}
\alias{interlabstats.formula}
\alias{interlabstats.default}
\alias{print.interlab}

\title{
Statistics for interlaboratory studies
}
\description{
  Calculates statistics for interlaboratory studies.
  \code{interlabstats} is a generic function.
  
  \code{interlabstats.formula} generates the statistics based on a
  simple analysis of variance.

  \code{interlabstats.default} uses a suitable fitted model object,
  currently either an \code{lmeMod} or a \code{rlmeMod} object,
  and extracts the desired statistics from it.
}
\usage{
interlabstats(object, ...)
\method{interlabstats}{default}(object)
\method{interlabstats}{formula}{formula, data, subset}
\method{print}{interlab}(x, groups=TRUE, ...)
}
\arguments{
  \item{object}{for \code{interlabstats.default}: object of class
    \code{lmerMod} or \code{rlmerMod}, obtained from function
    \code{lmer} of package \code{lme4} or \code{rlmer} of package
    \code{robustlmm}, respectively.}
  \item{formula}{
    Formula of type  \code{y ~ group} , where  \code{y}  is the
    measurement and \code{group}  is the laboraotory factor}
  \item{data}{dataset containing the variables used in \code{formula}}
  \item{subset}{usual subset argument, passed to \code{model.frame}}
  \item{groups}{for \code{print.interlab}: should statistics for groups
    be printed (if available)?}
}
\details{
  By \code{interlab.formula}, the estimated quantities are obtained from
  a classical fixed effects Analysis of Variance.
}
\value{A list containig
  \item{mean}{overall mean}
  \item{sigma}{within-group standard deviation}
  \item{siggroup}{between-group standard deviation}
  \item{repeatability}{standard error of a difference between
    measurements of the same lab}
  \item{reproducibility}{same, for different labs}
  \item{groups}{table, containing number of observations, mean and
    standard deviation for each group}
  \item{data}{the two relevant variables of the original data.frame}
  \item{call}{the call to the function}
  \item{method}{estimatrion method.
    "classical" for \code{interlabstats.formula},
    \code{class(object)} otherwise.}
  Argumentx of \code{print.interlab}:
  \item{x}{interlab object, as generated by \code{interlabstats}}
  \item{groups}{Should group table be printed?}
  \item{...}{Further arguments passed to \code{format}}
}
\references{
Stahel ...
}
\author{{Werner A. Stahel, ETH Zurich}
\seealso{\code{\link{plot.interlab}, \code{\link{lmer}}, \code{\link{rlmer}} 
%%}
\examples{
data(d.perm)
dd <- d.perm[d.perm$section=="w.1",]
dd[dd$team=="A","perm.log"] <- dd[dd$team=="A","perm.log"] - 0.3
  ## otherwise, there is no between group variation
( r.cl <- interlabstats(perm.log~team, data=dd) )
print(r.cl, groups=FALSE)

## Mixed models: lme
require(lme4)
r.lme <- lmer(perm.log~(1|team), data=dd, na.action=na.omit)
summary(r.lme)
interlabstats(r.lme)

## robust
require(robustlmm)
r.rlme <- rlmer(perm.log~(1|team), data=dd, na.action=na.omit)
summary(r.rlme)
interlabstats(r.rlme)
}
% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory.
\keyword{quality control}
\keyword{anova}% __ONLY ONE__ keyword per line
